feat: 약속 장소 검색 결과 칸수 고정(빈칸 포함)
#43
